feat: result_play_info 를 가진 playerInfo 들을 찾는 기능 구현
